requirements.txt
# Core
langgraph>=0.3.0
langchain-core>=0.3.0

# LLM Providers (tiered)
langchain-groq>=0.2.0
langchain-anthropic>=0.3.0
openai>=1.0.0

# RAG
chromadb>=0.5.0
langchain-chroma>=0.2.0

# Image Generation
httpx>=0.27.0
Pillow>=10.0.0

# API & Dashboard
fastapi>=0.115.0
uvicorn>=0.32.0
streamlit>=1.40.0

# Utilities
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
